+++ CHANGELOG.md
@@ -2,16 +2,16 @@
 
 All notable changes to `EchoQuery` will be documented in this file.
 
-## Last update 07-04-2024
+## Last update 08-04-2024
 
 ### Added
-- Where statement comparison operator refactor
+- Where statement logical operator refactor
 
 ### Deprecated
 - Nothing
 
 ### Fixed
-- Nothing
+- Where statement more than one comparison operator switching comparisons fixed
 
 ### Removed
 - Nothing